Consolidated unit tests to minimize docker containers and speed up test execution.
