+++ directoryServer/src/main/resources/templates/documentation.html
@@ -244,10 +244,9 @@
                 <li><a href="#create-serviceable-institution">Create a serviceable institution</a></li>
                 <li>Updating organization information in the directory</li>
             </ol>
-            Code for these tasks can be seen in the network server's implemenation itself, or in the unit tests.  The unit
-            test classes have been named for easy identification of use cases.  For example, service provider tests are
-            located in the "ServiceProviderUseCaseTests" and institution tests are in the "InstitutionUseCaseTests" and
-            digital signature tests are in the "DigitalSignatureTests" Java classes.
+            Code for these tasks can be seen in the network server's implemenation itself, or in the unit tests found in either
+            the network server or the directory server or both. The unit test names and comments indicate what organization type (institution or
+            service provider) the test case is designed to simulate.
             <p>
                 It is likely that more sample code in the form of unit tests will be added to the network server project in the future.
             </p>
