+++ networkServer/src/test/java/org/pesc/DigitalSignatureTests.java
@@ -6,6 +6,7 @@ import org.junit.runner.RunWith;
 import org.pesc.cds.service.PKIService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
@@ -17,6 +18,7 @@ import static org.junit.Assert.assertTrue;
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = NetworkServerApplication.class)
 @WebAppConfiguration
+@DirtiesContext
 public class DigitalSignatureTests {
 
 	@Autowired
